organize code related to parallel application of log recovery in a single header & source file - log_recovery_redo_parallel
