+++ cubrid/CMakeLists.txt
@@ -296,8 +296,7 @@ set(TRANSACTION_SOURCES
   ${TRANSACTION_DIR}/log_prior_recv.cpp
   ${TRANSACTION_DIR}/log_reader.cpp
   ${TRANSACTION_DIR}/log_recovery.c
-  ${TRANSACTION_DIR}/log_recovery_redo_bucket_queue.cpp
-  ${TRANSACTION_DIR}/log_recovery_redo_task.cpp
+  ${TRANSACTION_DIR}/log_recovery_redo_parallel.cpp
   ${TRANSACTION_DIR}/log_system_tran.cpp
   ${TRANSACTION_DIR}/log_tran_table.c
   ${TRANSACTION_DIR}/log_writer.c
@@ -322,7 +321,7 @@ set(TRANSACTION_HEADERS
   ${TRANSACTION_DIR}/log_prior_recv.hpp
   ${TRANSACTION_DIR}/log_reader.hpp
   ${TRANSACTION_DIR}/log_recovery_redo_log_rec.hpp
-  ${TRANSACTION_DIR}/log_recovery_redo.hpp
+  ${TRANSACTION_DIR}/log_recovery_redo_parallel.hpp
   ${TRANSACTION_DIR}/log_record.hpp
   ${TRANSACTION_DIR}/log_storage.hpp
   ${TRANSACTION_DIR}/log_system_tran.hpp
