+++ unit_tests/recovery/main.cpp
@@ -1,5 +1,5 @@
 
-#include "log_recovery_redo.hpp"
+//#include "log_recovery_redo.hpp"
 //#include "log_recovery_redo_debug_helpers.hpp"
 #include "storage_common.h"
 #include "thread_compat.hpp"
